requirements.txt
pandas
numpy
scikit-learn
flask
python-dotenv
google-generativeai
joblib
boto3